requirements.txt

argparse==1.4.0
asgi-redis==0.12.0
asgiref==0.13.0
autobahn==0.14.1
channels==0.14.0
contextlib2==0.5.3
daphne==0.12.1
dj-database-url==0.4.1
django-configurations==1.0
django-extensions==1.6.7
django-model-utils==2.5
django-stdimage==2.3.3
Django==1.9.6
msgpack-python==0.4.7
pillow==3.2.0
progressbar2==3.6.2
psycopg2==2.6.1
raven==5.19.0
redis==2.10.5
six==1.10.0
twisted==16.2.0
txaio==2.5.1
uWSGI==2.0.13.1
whitenoise==3.2
zope.interface==4.1.3

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools
